refactor: split retrieveQuestion and queue
