+++ src/utility/queue.ts
@@ -1,0 +1,5 @@
+import { Tweet } from "../interfaces/twitter/tweet.interface";
+
+export const queueQuestion: Tweet[] = [];
+export const queueSummary: any[] = [];
+export const queueReport: any[] = [];
